separate AuthZ Script into its own class
