+++ plantain/src/main/scala/plantain/PlantainStore.scala
@@ -294,7 +294,6 @@ class PlantainLDPCActor(baseUri: URI, root: Path) extends Actor {
 
 }
 
-
 object PlantainLDPS {
 
   case class CreateLDPC(uri: URI)
